djangoproj/requirements.txt
django
django-cors-headers
whitenoise
django-ninja
daphne
python-dotenv
python-jose
google-cloud-pubsub
requests
google-auth
pusher
